Revert "add backend tests (#84)"
This reverts commit dff8b09618e58f323f8508c86246f4396a1c13f0.
