+++ .github/workflows/build.yaml
@@ -8,7 +8,7 @@ jobs:
     defaults:
       run:
         working-directory: backend
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-24.04
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
@@ -19,14 +19,14 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Run build
+      - name: Build project
         run: npm run build
 
   frontend:
     defaults:
       run:
         working-directory: frontend
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-24.04
     steps:
       - name: Checkout code
         uses: actions/checkout@v4
@@ -37,5 +37,5 @@ jobs:
       - name: Install dependencies
         run: npm ci
 
-      - name: Run build
+      - name: Build project
         run: npm run build
