--- .github/workflows/test.yaml
@@ -1,23 +1,0 @@
-name: Test
-
-on:
-  workflow_call:
-
-jobs:
-  backend:
-    defaults:
-      run:
-        working-directory: backend
-    runs-on: ubuntu-latest
-    steps:
-      - name: Checkout code
-        uses: actions/checkout@v4
-
-      - name: Setup Node
-        uses: actions/setup-node@v4
-
-      - name: Install dependencies
-        run: npm ci
-
-      - name: Run test
-        run: npm run test
