+++ backend/handlers/exchangeActions.ts
@@ -1,30 +1,27 @@
-import { pool } from '../db/pool';
-import { getEnvValue } from '../utils/getEnv';
-import { sendPushNotification } from '../utils/sendPushNotification';
-
-import jwt from 'jsonwebtoken';
+import { pool } from '../db/pool.js';
 import { Request, Response, NextFunction } from 'express';
+import jwt from 'jsonwebtoken';
+import { getEnvValue } from '../utils/getEnv.js';
+import { sendPushNotification } from '../utils/sendPushNotification.js';
 
 const JWT_KEY = getEnvValue('JWT_KEY', 'JWT_KEY_FILE')!;
 
 export const completeExchangeHandler = async (req: Request, res: Response, _next: NextFunction) => {
     try {
         const connectedUsers: Map<string, string> = req.app.get('connectedUsers');
+        
         const authHeader = req.headers.authorization;
         if (!authHeader || !authHeader.startsWith('Bearer ')) {
-            res.status(401).json({ message: 'Token manquant.' });
-
-            return;
+            return res.status(401).json({ message: 'Token manquant.' });
         }
 
         const token = authHeader.split(' ')[1];
         const decoded = jwt.verify(token, JWT_KEY) as { id: number, email: string };
         const userId = decoded.id;
+
         const exchangeId = parseInt(req.params.id);
         if (!exchangeId) {
-            res.status(400).json({ message: 'ID de l\'échange requis.' });
-
-            return;
+            return res.status(400).json({ message: 'ID de l\'échange requis.' });
         }
 
         const exchangeResult = await pool.query(
@@ -32,26 +29,20 @@ export const completeExchangeHandler = async (req: Request, res: Response, _next
             [exchangeId]
         );
         if (exchangeResult.rows.length === 0) {
-            res.status(404).json({ message: 'Échange non trouvé.' });
-
-            return;
+            return res.status(404).json({ message: 'Échange non trouvé.' });
         }
 
         const exchange = exchangeResult.rows[0];
         if (exchange.initiator_id !== userId && exchange.recipient_id !== userId) {
-            res.status(403).json({
+            return res.status(403).json({
                 message: 'Vous n\'êtes pas autorisé à terminer cet échange.'
             });
-
-            return;
         }
 
         if (exchange.status !== 'confirmed') {
-            res.status(400).json({
+            return res.status(400).json({
                 message: 'L\'échange doit être confirmé par les deux parties avant d\'être terminé.'
             });
-
-            return;
         }
 
         await pool.query(
@@ -112,14 +103,13 @@ export const getUserExchangesHandler = async (req: Request, res: Response, _next
     try {
         const authHeader = req.headers.authorization;
         if (!authHeader || !authHeader.startsWith('Bearer ')) {
-            res.status(401).json({ message: 'Token manquant.' });
-
-            return;
+            return res.status(401).json({ message: 'Token manquant.' });
         }
 
         const token = authHeader.split(' ')[1];
         const decoded = jwt.verify(token, JWT_KEY) as { id: number, email: string };
         const userId = decoded.id;
+
         const exchangesQuery = `
             SELECT 
                 e.*,
@@ -137,6 +127,7 @@ export const getUserExchangesHandler = async (req: Request, res: Response, _next
             WHERE e.initiator_id = $1 OR e.recipient_id = $1
             ORDER BY e.created_at DESC
         `;
+        
         const result = await pool.query(exchangesQuery, [userId]);
         const exchanges = result.rows.map(row => ({
             id: row.id,
@@ -176,29 +167,25 @@ export const getExchangeRatingHandler = async (req: Request, res: Response, _nex
     try {
         const authHeader = req.headers.authorization;
         if (!authHeader || !authHeader.startsWith('Bearer ')) {
-            res.status(401).json({ message: 'Token manquant.' });
-
-            return;
+            return res.status(401).json({ message: 'Token manquant.' });
         }
 
         const token = authHeader.split(' ')[1];
         const decoded = jwt.verify(token, JWT_KEY) as { id: number, email: string };
         const userId = decoded.id;
+
         const exchangeId = parseInt(req.params.id);
         if (!exchangeId) {
-            res.status(400).json({ message: 'ID de l\'échange requis.' });
-
-            return;
+            return res.status(400).json({ message: 'ID de l\'échange requis.' });
         }
 
         const result = await pool.query(
             `SELECT * FROM profile_ratings WHERE exchange_id = $1 AND rater_id = $2`,
             [exchangeId, userId]
         );
+        
         if (result.rows.length === 0) {
-            res.json(null);
-
-            return;
+            return res.json(null);
         }
 
         res.json(result.rows[0]);
