+++ backend/handlers/pushNotifications.ts
@@ -1,8 +1,7 @@
-import { pool } from '../db/pool';
-import { getEnvValue } from '../utils/getEnv';
-
+import { pool } from '../db/pool.js';
 import { Request, Response, NextFunction } from 'express';
 import jwt from 'jsonwebtoken';
+import { getEnvValue } from '../utils/getEnv.js';
 
 const JWT_KEY = getEnvValue('JWT_KEY', 'JWT_KEY_FILE')!;
 
@@ -10,11 +9,8 @@ export const getVapidPublicKeyHandler = async (req: Request, res: Response, _nex
     try {
         const vapidPublicKey = process.env.VAPID_PUBLIC_KEY;
         if (!vapidPublicKey) {
-            res.status(500).json({ message: 'Clé VAPID non configurée.' });
-
-            return;
+            return res.status(500).json({ message: 'Clé VAPID non configurée.' });
         }
-
         res.send(vapidPublicKey);
     } catch (err) {
         console.error('Erreur lors de la récupération de la clé VAPID:', err);
@@ -28,19 +24,16 @@ export const subscribeToPushHandler = async (req: Request, res: Response, _next:
     try {
         const authHeader = req.headers.authorization;
         if (!authHeader || !authHeader.startsWith('Bearer ')) {
-            res.status(401).json({ message: 'Token manquant.' });
-
-            return;
+            return res.status(401).json({ message: 'Token manquant.' });
         }
 
         const token = authHeader.split(' ')[1];
         const decoded = jwt.verify(token, JWT_KEY) as { id: number, email: string };
         const userId = decoded.id;
+
         const { subscription } = req.body;
         if (!subscription) {
-            res.status(400).json({ message: 'Subscription requise.' });
-
-            return;
+            return res.status(400).json({ message: 'Subscription requise.' });
         }
 
         const userCheck = await pool.query(
@@ -48,9 +41,7 @@ export const subscribeToPushHandler = async (req: Request, res: Response, _next:
             [userId]
         );
         if (userCheck.rows.length === 0) {
-            res.status(404).json({ message: 'Utilisateur non trouvé ou compte désactivé.' });
-
-            return;
+            return res.status(404).json({ message: 'Utilisateur non trouvé ou compte désactivé.' });
         }
 
         await pool.query(
@@ -64,6 +55,7 @@ export const subscribeToPushHandler = async (req: Request, res: Response, _next:
             [userId, JSON.stringify(subscription)]
         );
 
+        console.log(`✅ Utilisateur ${userId} abonné aux notifications push`);
         res.json({ message: 'Abonnement aux notifications réussi.' });
     } catch (err) {
         console.error('Erreur lors de l\'abonnement aux notifications:', err);
@@ -77,9 +69,7 @@ export const unsubscribeFromPushHandler = async (req: Request, res: Response, _n
     try {
         const authHeader = req.headers.authorization;
         if (!authHeader || !authHeader.startsWith('Bearer ')) {
-            res.status(401).json({ message: 'Token manquant.' });
-
-            return;
+            return res.status(401).json({ message: 'Token manquant.' });
         }
 
         const token = authHeader.split(' ')[1];
@@ -91,6 +81,7 @@ export const unsubscribeFromPushHandler = async (req: Request, res: Response, _n
             [userId]
         );
 
+        console.log(`❌ Utilisateur ${userId} désabonné des notifications push`);
         res.json({ message: 'Désabonnement des notifications réussi.' });
     } catch (err) {
         console.error('Erreur lors du désabonnement des notifications:', err);
