--- backend/jest.config.ts
@@ -1,11 +1,0 @@
-import type { Config } from 'jest';
-
-const config: Config = {
-    preset: 'ts-jest',
-    testEnvironment: 'node',
-    testMatch: ['**/tests/**/*.test.ts'],
-    moduleFileExtensions: ['ts', 'js'],
-    roots: ['<rootDir>']
-};
-
-export default config;
