+++ backend/jobs/deleteUnsubscribedUsers.ts
@@ -1,4 +1,4 @@
-import { pool } from "../db/pool";
+import { pool } from "../db/pool.js";
 
 export const deleteUnsubcribedUsers = async () => {
     try {
