+++ backend/package.json
@@ -1,46 +1,33 @@
 {
-    "type": "commonjs",
-    "main": "src/index.ts",
-    "scripts": {
-        "build": "tsc",
-        "start": "node build/index.js",
-        "dev": "node --loader ts-node/esm index.ts",
-        "test": "jest"
-    },
-    "dependencies": {
-        "bcrypt": "^6.0.0",
-        "cors": "^2.8.5",
-        "date-fns": "^4.1.0",
-        "dotenv": "^16.5.0",
-        "express": "^5.1.0",
-        "jsonwebtoken": "^9.0.2",
-        "node-cron": "^4.1.1",
-        "pg": "^8.16.0",
-        "socket.io": "^4.8.1",
-        "typescript": "^5.8.3",
-        "web-push": "^3.6.7"
-    },
-    "devDependencies": {
-        "@types/bcrypt": "^5.0.2",
-        "@types/cors": "^2.8.18",
-        "@types/dotenv": "^6.1.1",
-        "@types/express": "^5.0.1",
-        "@types/jest": "^30.0.0",
-        "@types/jsonwebtoken": "^9.0.9",
-        "@types/node": "^22.15.18",
-        "@types/pg": "^8.15.2",
-        "@types/socket.io": "^3.0.1",
-        "@types/supertest": "^6.0.3",
-        "@types/web-push": "^3.6.4",
-        "jest": "^30.0.4",
-        "supertest": "^7.1.3",
-        "ts-jest": "^29.4.0",
-        "ts-node": "^10.9.2"
-    },
-    "overrides": {
-        "glob": "^9.0.0",
-        "some-package": {
-            "inflight": "npm:lru-cache@^10.0.0"
-        }
-    }
+  "scripts": {
+    "build": "tsc",
+    "start": "node build/index.js",
+    "dev": "node --loader ts-node/esm index.ts"
+  },
+  "dependencies": {
+    "bcrypt": "^6.0.0",
+    "cors": "^2.8.5",
+    "date-fns": "^4.1.0",
+    "dotenv": "^16.5.0",
+    "express": "^5.1.0",
+    "jsonwebtoken": "^9.0.2",
+    "node-cron": "^4.1.1",
+    "pg": "^8.16.0",
+    "socket.io": "^4.8.1",
+    "typescript": "^5.8.3",
+    "web-push": "^3.6.7"
+  },
+  "type": "module",
+  "devDependencies": {
+    "@types/bcrypt": "^5.0.2",
+    "@types/cors": "^2.8.18",
+    "@types/dotenv": "^6.1.1",
+    "@types/express": "^5.0.1",
+    "@types/jsonwebtoken": "^9.0.9",
+    "@types/node": "^22.15.18",
+    "@types/pg": "^8.15.2",
+    "@types/socket.io": "^3.0.1",
+    "@types/web-push": "^3.6.4",
+    "ts-node": "^10.9.2"
+  }
 }
