+++ backend/tsconfig.json
@@ -1,23 +1,15 @@
 {
-    "compilerOptions": {
-        "target": "es2016",
-        "module": "commonjs",
-        "esModuleInterop": true,
-        "forceConsistentCasingInFileNames": true,
-        "strict": true,
-        "outDir": "build",
-        "skipLibCheck": true,
-        "types": [
-            "@types/express",
-            "jest"
-        ]
-    },
-    "include": [
-        "src"
-    ],
-    "exclude": [
-        "tests",
-        "**/*.test.ts",
-        "jest*"
-    ]
+  "compilerOptions": {
+    "target": "ESNext",
+    "module": "NodeNext",
+    "moduleResolution": "nodenext",
+    "outDir": "build",
+    "noEmitOnError": true,
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "strict": true,
+    "skipLibCheck": true,
+    "noImplicitAny": true,
+    "types": ["@types/express"]
+  }
 }
