+++ backend/utils/origins.ts
@@ -1,4 +1,4 @@
-import { getEnvListValue } from "./getEnv";
+import { getEnvListValue } from "./getEnv.js";
 
 export function getAllowedOrigins(): string[] {
     return getEnvListValue('ALLOWED_ORIGINS', 'ALLOWED_ORIGINS_FILE', ['http://localhost:8080']);
